requirements.txt
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
six==1.17.0
